♻️ Split ports from expansion policy (Task_2)
Port traits move to src/ports/; the ADR-I-0006 bounded-expansion
algorithm unifies in src/policy/graph_expansion.rs, deleting the
Oxigraph adapter's duplicated helper copies. The adapter-side link-ref
flavor is semantically distinct (pre-hydration pruning) and is
colocated with a documented difference rather than force-merged.
internal/repositories.rs remains as a compatibility barrel until
pipelines move in Task_3.
