+++ src/internal/repositories.rs
@@ -1,22 +1,18 @@
 mod correction_forget_pipeline;
-mod embedder;
-mod graph_authority_store;
 mod link_pipeline;
 mod reconciliation;
 mod remember_pipeline;
 mod retrieval_selectivity;
-mod retrieval_stats_store;
 mod retrieve_pipeline;
-mod source_reference;
 #[cfg(test)]
 pub(crate) mod test_support;
-mod vector_candidate_store;
 mod write_planning;
 
-// Internal contract surface. Pipelines, adapters, and test support consume
-// different subsets, so keep the module boundary stable.
+// Compatibility re-exports: port traits moved to crate::ports and the
+// bounded-expansion algorithm moved to crate::policy. Pipelines keep
+// importing through this barrel until they move in Task_3.
 #[allow(unused_imports)]
-pub(crate) use embedder::MemoryEmbedder;
+pub(crate) use crate::ports::embedder::MemoryEmbedder;
 
 // Lifecycle pipeline surface.
 #[allow(unused_imports)]
@@ -25,9 +21,12 @@ pub(crate) use correction_forget_pipeline::CorrectionForgetPipeline;
 // Graph-authority contract surface. Retrieval and lifecycle code use
 // different subsets of the query/expansion helpers.
 #[allow(unused_imports)]
-pub(crate) use graph_authority_store::{
+pub(crate) use crate::policy::graph_expansion::{
     apply_fanout_limits_by_pair, bounded_expansion, bounded_expansion_node_set,
     bounded_hub_retention_limit, derived_memories_by_provenance, derived_memories_by_thread,
+};
+#[allow(unused_imports)]
+pub(crate) use crate::ports::graph_authority::{
     GraphAuthorityStore, GraphDerivedMemoryProvenanceQuery, GraphDerivedMemoryThreadQuery,
     GraphExpansion, GraphExpansionBoundedFailure, GraphExpansionBoundedFailureReason,
     GraphExpansionFailurePolicy, GraphExpansionFanoutOverride, GraphExpansionFilteredNode,
@@ -43,7 +42,7 @@ pub(crate) use link_pipeline::{
 // Source-reference contracts remain internal; core stores opaque references,
 // not caller-owned source material.
 #[allow(unused_imports)]
-pub(crate) use source_reference::{ResolvedSourceReference, SourceReferenceResolver};
+pub(crate) use crate::ports::source_reference::{ResolvedSourceReference, SourceReferenceResolver};
 
 // Internal/admin reconciliation diagnostics. These remain out of the public
 // CharacterMemory facade until a governance surface is planned.
@@ -62,9 +61,9 @@ pub(crate) use remember_pipeline::{
 };
 
 #[cfg(test)]
-pub(crate) use retrieval_stats_store::noop_retrieval_stats_store;
+pub(crate) use crate::ports::retrieval_stats::noop_retrieval_stats_store;
 #[allow(unused_imports)]
-pub(crate) use retrieval_stats_store::{
+pub(crate) use crate::ports::retrieval_stats::{
     object_type_key, record_stats_after_write, relation_type_key, retention_state_key,
     retrieval_stats_edges, retrieval_stats_object_states, InMemoryRetrievalStatsStore,
     RetrievalStatsCounter, RetrievalStatsCounterKey, RetrievalStatsEdge, RetrievalStatsHealth,
@@ -83,7 +82,7 @@ pub(crate) use retrieve_pipeline::RetrievePipeline;
 
 // Vector candidate recall contract surface.
 #[allow(unused_imports)]
-pub(crate) use vector_candidate_store::VectorCandidateStore;
+pub(crate) use crate::ports::vector_candidate::VectorCandidateStore;
 
 #[allow(unused_imports)]
 pub(crate) use write_planning::{
