+++ src/lib.rs
@@ -5,6 +5,8 @@ mod errors;
 
 pub mod api;
 mod internal;
+pub(crate) mod policy;
+pub(crate) mod ports;
 
 use async_trait::async_trait;
