+++ src/policy.rs
@@ -1,0 +1,4 @@
+// Retrieval/expansion policies: provider-neutral algorithms that implement
+// core guarantees (e.g. ADR-I-0006 bounded graph expansion) independently of
+// any storage backend.
+pub(crate) mod graph_expansion;
