+++ src/ports/retrieval_stats.rs
