+++ src/ports/vector_candidate.rs
